zero grad

+++ negative_learning.py
@@ -108,7 +108,7 @@ print('Warm up training')
 trainset.active = 0
 dataloader = DataLoader(trainset, batch_size=args.batch_size, shuffle=True, num_workers=4)
 for i_batch, sample in enumerate(tqdm(dataloader)):
-    image = sample['image'].cuda()
+    image = sample['image'].float().cuda()
 
     reconstruction = ae(image)
